+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.18.1
  - bug fixes in `LogNormalFitter` variance estimates
- - improve convergence of `LogNormalFitter`
+ - improve convergence of `LogNormalFitter`. We now model the log of sigma internally, but still expose sigma externally. 
  - use the `autograd` lib to help with gradients
 
 ### 0.18.0